# WebAPI/bin/Debug/net6.0/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Web.Api",
    "version": "v1"
  },
  "paths": {
...
    "/api/Category/GetCategory": {
      "get": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Category/InsertCategory": {
      "post": {
        "tags": [
          "Category"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInsertModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInsertModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInsertModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Category/UpdateCategory": {
      "put": {
        "tags": [
          "Category"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdateModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdateModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdateModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Category/DeleteCategory": {
      "delete": {
        "tags": [
          "Category"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/UserType/GetUserType": {
      "get": {
        "tags": [
          "UserType"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserTypeViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTypeViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTypeViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/UserType/InsertUserType": {
      "post": {
        "tags": [
          "UserType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTypeInsertModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTypeInsertModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserTypeInsertModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/UserType/UpdateUserType": {
      "put": {
        "tags": [
          "UserType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTypeUpdateModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTypeUpdateModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserTypeUpdateModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/UserType/DeleteUserType": {
      "delete": {
        "tags": [
          "UserType"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CategoryInsertModel": {
        "required": [
          "categoryName"
        ],
        "type": "object",
        "properties": {
          "categoryName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CategoryUpdateModel": {
        "required": [
          "categoryName",
          "id"
        ],
        "type": "object",
        "properties": {
          "categoryName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "UserTypeViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "typeName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserTypeInsertModel": {
        "required": [
          "typeName"
        ],
        "type": "object",
        "properties": {
          "typeName": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UserTypeUpdateModel": {
        "required": [
          "id",
          "typeName"
        ],
        "type": "object",
        "properties": {
          "typeName": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 1safsfsdfdfd\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}